+++ zomes/social_context/src/utils.rs
@@ -1,0 +1,19 @@
+use hdk::prelude::*;
+use std::hash::Hash;
+use chrono::{Utc, DateTime, NaiveDateTime};
+
+use crate::errors::SocialContextResult;
+
+pub fn get_now() -> SocialContextResult<DateTime<Utc>> {
+    let now = sys_time()?.as_seconds_and_nanos();
+    Ok(DateTime::<Utc>::from_utc(
+        NaiveDateTime::from_timestamp(now.0, now.1),
+        Utc,
+    ))
+}
+
+pub fn dedup<T: Eq + Hash + Clone>(vs: &Vec<T>) -> Vec<T> {
+    let hs = vs.iter().cloned().collect::<HashSet<T>>();
+
+    hs.into_iter().collect()
+}
